fix: race condition that causes panics in core transaction manager
There was a race in the TSM between the Put and Send methods, between Send releasing the lock (but keeping a reference to the state) and Put closing the states chan. Sending on a closed chan causes a panic!

The solution feels more complicated than it should be, but it works and the new test passes.