feat: much better error handling for read requests
This change introduces types.Error and types.PropertyAccessError which are now used whenever the remote device sends us an error response.

The majority of the code is around ReadMultipleProperty support as the remote device can return an error for individual property reads which both wasn't supported by the go api and was being handled badly by the code. Before it would bail at the first property error, potentially skipping processing responses. Now it processes and returns the entire response but returns a multierr containing each individual failure in case the caller wants to use it for optimisation.

Utilities that read multiple properties have also been updated to pay attention to the new error type.